Module riscv_vm::devices

Devices

Devices are the main way of extending vm behavior and adding interfaces.

There devices come in two forms, handled devices and async devices, handled devices are updated each clock cycle and are intended for simple short jobs while async devices run independently of the main thread/clock and are intended for anything else, they have to request their next update by returning a AsyncDeviceUpdateResult.

Memory

Devices are permitted to add memory regions to the vm’s memory, the behavour of this memory is completely up to the device, with the only requirement being that these regions implement MemoryBuffer.

Modules

async_device

handled_device

memory_buffer

simple_uart

Structs

DeviceMemHandle

Gives devices access to memory during their initialization, allows for the registering of device memory regions.

Enums

DeviceError

DeviceInitError

Traits

Device

Part one of the trifecta of traits that make up a device, defines the size of memory shared with the vm and the Device::new() function, which defines how to create (but not initialize) the device, essentially the pre poweron state of the device.

DeviceObject

Part two of the trifecta of traits that make up a device. The init functions is ran when the vm and devices are initializing but before any code is ran, this is the time when the device can register any memory regions, this initialization is allowed to error, these errors should be passed up as DeviceInitError::Other.