Trait riscv_vm::devices::handled_device::HandledDevice

pub trait HandledDevice: Debug + DeviceObject {
    // Required methods
    fn update(
        &mut self,
        mem: &mut DeviceMemory,
        event_bus: &DeviceEventBusHandle,
        data: DeviceData
    ) -> Result<(), DeviceError>;
    fn event(
        &mut self,
        mem: &mut DeviceMemory,
        event: DeviceEvent,
        event_bus: &DeviceEventBusHandle
    ) -> Result<(), DeviceError>;
}

Part three of a handled device, this trait defines the behaviour of a handled device. These devices run in sync with the main clock and are meant for simple, short, tasks.

Required Methods

fn update( &mut self, mem: &mut DeviceMemory, event_bus: &DeviceEventBusHandle, data: DeviceData ) -> Result<(), DeviceError>

The main way for the device to do logic, this function is called once per clock cycle.

fn event( &mut self, mem: &mut DeviceMemory, event: DeviceEvent, event_bus: &DeviceEventBusHandle ) -> Result<(), DeviceError>

Called when an event occurs which the device needs to be notified of, the device may choose to ignore this.