Trait riscv_vm::devices::memory_buffer::MemoryBuffer

pub trait MemoryBuffer {
    // Required methods
    fn size(&self) -> u64;
    fn write_bytes(
        &mut self,
        bytes: &[u8],
        addr: Address
    ) -> Result<(), MemoryBufferError>;
    fn read_bytes(
        &self,
        addr: Address,
        size: usize
    ) -> Result<Vec<u8>, MemoryBufferError>;
}

A buffer mapped to a specific address in memory, to which the vm can read and write there does not have to be a 1:1 backing in host memory, a read or write call might convert data stored in a different way to u8 or might create any data only when called.

All addr arguments are normalised so that 0x0 is at the base of the buffer

Required Methods

fn size(&self) -> u64

fn write_bytes( &mut self, bytes: &[u8], addr: Address ) -> Result<(), MemoryBufferError>

fn read_bytes( &self, addr: Address, size: usize ) -> Result<Vec<u8>, MemoryBufferError>