Trait riscv_vm::devices::Device

pub trait Device {
    const MEM_SIZE: u64;

    // Required method
    fn new() -> Self;
}

Part one of the trifecta of traits that make up a device, defines the size of memory shared with the vm and the Device::new() function, which defines how to create (but not initialize) the device, essentially the pre poweron state of the device.

Required Associated Constants

const MEM_SIZE: u64

Required Methods

fn new() -> Self

Object Safety

This trait is not object safe.