Module riscv_vm::vmstate

The vmstate is the main way to interact with the vm, is is created via a VMStateBuilder and can than be interacted with directly.

Structs

VMSettings

VMState

An actual instance of a riscv vm, with memory, devices and harts

VMStateBuilder

Used to setup a VM, create a builder using Default::default() or VMStateBuilder::new(), see the methods below for available options, use VMStateBuilder::build() to turn into a usable VMState.

Enums

KernelLoadError

VMError

VMInitError