Struct riscv_vm::vmstate::VMState

pub struct VMState { /* private fields */ }

An actual instance of a riscv vm, with memory, devices and harts

Implementations

impl VMState

pub fn load_elf_kernel(&mut self, elf: &Elf) -> Result<(), VMError>

Load a kernel from an elf file and place it at 0x80000000 (bottom of memory) The elf must be riscv64 little endian.

pub fn step(&mut self, verbose: bool) -> Result<(), VMError>

Advance all cores one cycle and, if verbose, print the instruction that was executed

pub fn step_hart_until( &mut self, hart: usize, target: Address ) -> Result<(), VMError>

Step a specific hart until its pc hits the given address or it has made 10000 steps, whichever happens first

pub fn step_all_until(&mut self, target: Address) -> Result<(), VMError>

Step all harts until its pc hits the given address or it has made 10000 steps, whichever happens first

pub fn run(&mut self) -> Result<(), VMError>

Run the vm until it errors or forever, whichever happens first.

pub fn fetch(&mut self, hart: usize) -> Result<Instruction, MemoryError>

Attempt to fetch on a specific hart and return the decoded instruction

pub fn dump_mem(&self)

๐Deprecated

pub fn print_mem_map(&self)

๐Deprecated

pub fn get_hart(&self, hart: usize) -> Option<&Hart>

Get an immutable refierence to a specific hart, if it exists.