Module riscv_vm::devices

Devices

Devices are the main way of extending vm behavior and adding interfaces.

There devices come in two forms, handled devices and async devices, handled devices are updated each clock cycle and are intended for simple short jobs while async devices run independently of the main thread/clock and are intended for anything else.

Data Storage

Devices are offered 3 forms of data storage. The first is the object that implements the device traits, this form of data storage is internal to the device and not accessible to the vm. Second there is DeviceMemory, this is memory shared between the device and the vm, the size of this memory is requested by setting the MEM_SIZE constant in the Device trait, the location of this memory in the device’s memory map is determined when instantating the device, this memory may allow data races and non exclusive write access, devices may define their own ways of communicating with the vm would they want to avoid this. The third form of data is available as DeviceData returned when initializing a device, this data is given back to the device on update and is also given to poll memory registers, not that this data is given as a reference to a Box<dyn Any>, the user is responsible for casting to the right type.

Modules

async_device

event_bus

handled_device

simple_uart

Structs

DeviceMemory

MemoryRegisterHandle

Enums

DeviceError

DeviceInitError

Register

Traits

Device

Part one of the trifecta of traits that make up a device, defines the size of memory shared with the vm and the Device::new() function, which defines how to create (but not initialize) the device, essentially the pre poweron state of the device.

DeviceObject

Part two of the trifecta of traits that make up a device. The init functions is ran when the vm and devices are initializing but before any code is ran, this is the time when the device can register any memory mapped registers and return sync data to be shared with its registers, this initialization is allowed to error, these errors should be passed up as DeviceInitError::Other.

Type Aliases

DeviceData

Alias for the wrapper around the simple data shared between devices and their memory mapped registers, this type is given to the vm by the device upon init and is passed back to the devices and registers on any event, the [dyn Any + Send + Sync] is guarenteed to be of the same type as the data created by the device before being cast to an Any.